requirements.txt
streamlit
pandas
requests
google-generativeai>=0.8.3